+++ src/tool.tcl
@@ -784,7 +784,7 @@ proc tool::PrepareRunCommand {com fname} {
 }
 #_______________________
 
-proc tool::_run {{what ""} {runmode ""}} {
+proc tool::_run {{what ""} {runmode ""} args} {
   # Runs e_menu's item of menu.em.
   #   what - the item (by default, "Run me")
   #   runmode - mode of running (in console or in tkcon)
@@ -794,6 +794,12 @@ proc tool::_run {{what ""} {runmode ""}} {
   set opts "EX=$what"
   if {$what eq {}} {
     #  it is 'Run me' e_menu item
+    set doit [::apave::extractOptions args -doit 0]
+    lassign [alited::ExtTrans [alited::bar::FileName]] ext istrans from to
+    if {!$doit && $istrans} {
+      alited::hl_trans::translateLine
+      return
+    }
     if {!$::alited::DEBUG} {
       if {$al(EM,exec)} {
         set fpid [alited::TmpFile .pid~]